% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMapping-functions.R
\name{.transform_types}
\alias{.transform_types}
\title{Transform variable types according to field definition}
\usage{
.transform_types(table, fields)
}
\arguments{
\item{table}{}

\item{fields}{}
}
\value{

}
\description{
Transform variable types according to field definition
}
