% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-actions.R
\name{get_proto_action}
\alias{get_proto_action}
\title{Get a proto-action}
\usage{
get_proto_action(action, ...)
}
\arguments{
\item{action}{The action name}

\item{...}{Params}
}
\description{
Get an action ad-hoc from a parameter list for use in unit tests etc.
}
