% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-actions.R
\name{get_proto_backend}
\alias{get_proto_backend}
\title{Get proto-Backend object}
\usage{
get_proto_backend(variables = NULL, peaks = 4, ...)
}
\arguments{
\item{variables}{a tibble with variables set}

\item{peaks}{the number of peaks per spectrum}

\item{n}{the number of spectra to fake}
}
\description{
Get a minimally filled \code{MsBackendMapping} with some \code{variables} filled to run unittests etc.
}
