% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/MsBackendSchema.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{[,MsBackendSchema-method}
\alias{as.list,MsBackendSchema-method}
\alias{intensity,MsBackendSchema-method}
\alias{mz,MsBackendSchema-method}
\alias{lengths,MsBackendSchema-method}
\title{Internal page for hidden aliases}
\usage{
\S4method{as.list}{MsBackendSchema}(x)

\S4method{intensity}{MsBackendSchema}(object)

\S4method{mz}{MsBackendSchema}(object)

\S4method{lengths}{MsBackendSchema}(x, use.names = FALSE)
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
