% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-actions-default.R
\name{MetadataActionDefault}
\alias{MetadataActionDefault}
\title{Metadata action: default}
\description{
Metadata action: default

Metadata action: default
}
\details{
Sets a default value, on read and/or write, if no value is present for a particular column
Creates a column if no column exists.
Can be used on read (target), write (source) or both.
}
\examples{
backend <- get_proto_backend(
   readCol = list("POSITIVE", c(NA_character_), "n", "gugus", NA_character_),
   WRITE_COL = list("value", NA_character_, "value", "otherValue", c())
)
action <- get_proto_action(
   "default",
   params = list(
      c(target = 'default_new_read_col', read = 'new_read'),
      c(target = 'readCol', read = 'pos_default'),
      c(source = 'WRITE_COL', write = 'default'),
      c(source = 'WRITE_NEW_COL', write = 'one_value')

   )
)
fw <- action$execute_read(backend)
bw <- action$execute_write(fw)
}
\section{Super class}{
\code{\link[SpectraMapping:MetadataActionBase]{SpectraMapping::MetadataActionBase}} -> \code{MetadataActionDefault}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MetadataActionDefault-process_read}{\code{MetadataActionDefault$process_read()}}
\item \href{#method-MetadataActionDefault-process_write}{\code{MetadataActionDefault$process_write()}}
\item \href{#method-MetadataActionDefault-clone}{\code{MetadataActionDefault$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SpectraMapping" data-topic="MetadataActionBase" data-id="execute_read"><a href='../../SpectraMapping/html/MetadataActionBase.html#method-MetadataActionBase-execute_read'><code>SpectraMapping::MetadataActionBase$execute_read()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SpectraMapping" data-topic="MetadataActionBase" data-id="execute_write"><a href='../../SpectraMapping/html/MetadataActionBase.html#method-MetadataActionBase-execute_write'><code>SpectraMapping::MetadataActionBase$execute_write()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SpectraMapping" data-topic="MetadataActionBase" data-id="initialize"><a href='../../SpectraMapping/html/MetadataActionBase.html#method-MetadataActionBase-initialize'><code>SpectraMapping::MetadataActionBase$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SpectraMapping" data-topic="MetadataActionBase" data-id="log_level"><a href='../../SpectraMapping/html/MetadataActionBase.html#method-MetadataActionBase-log_level'><code>SpectraMapping::MetadataActionBase$log_level()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SpectraMapping" data-topic="MetadataActionBase" data-id="merge_settings"><a href='../../SpectraMapping/html/MetadataActionBase.html#method-MetadataActionBase-merge_settings'><code>SpectraMapping::MetadataActionBase$merge_settings()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SpectraMapping" data-topic="MetadataActionBase" data-id="set_settings"><a href='../../SpectraMapping/html/MetadataActionBase.html#method-MetadataActionBase-set_settings'><code>SpectraMapping::MetadataActionBase$set_settings()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataActionDefault-process_read"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataActionDefault-process_read}{}}}
\subsection{Method \code{process_read()}}{
read implementation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataActionDefault$process_read(data, params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataActionDefault-process_write"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataActionDefault-process_write}{}}}
\subsection{Method \code{process_write()}}{
write implementation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataActionDefault$process_write(data, params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataActionDefault-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataActionDefault-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataActionDefault$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
