% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMapping.R
\name{MsBackendMapping}
\alias{MsBackendMapping}
\alias{MsBackendMapping-class}
\alias{backendInitialize,MsBackendMapping-method}
\title{MS data backend for mapped text files}
\usage{
\S4method{backendInitialize}{MsBackendMapping}(object, files, data, nonStop = FALSE, parallel = FALSE, ...)

MsBackendMapping(format, fields = .load_default_fields())
}
\arguments{
\item{object}{Instance of \code{MsBackendMapping} class.}

\item{files}{\code{character} with the (full) file name(s) of the  file(s)
from which MS/MS data should be imported.}

\item{nonStop}{\code{logical(1)} whether import should be stopped if an
xml file does not contain all required fields. Defaults to
\code{nonStop = FALSE}.}

\item{...}{Currently ignored.}

\item{BPPARAM}{Parameter object defining the parallel processing
setup to import data in parallel. Defaults to \code{BPPARAM = bpparam()}. See \code{\link[=bpparam]{bpparam()}} for more information.}
}
\description{
The \code{MsBackendMapping} class supports import of MS/MS spectra data from
files in text format by a defined mapping.
After initial import, the full MS data is kept in
memory. \code{MsBackendMapping} uses tibble/tidyverse stuff in the background.

New objects are created with the \code{MsBackendMapping} function. The
\code{backendInitialize} method has to be subsequently called to
initialize the object and import MS/MS data from (one or more)
files.  Optional parameter \code{nonStop} allows to specify whether the
import returns with an error if one of the xml files lacks required
data, such as \code{mz} and \code{intensity} values (default \code{nonStop = FALSE}), or whether only affected file(s) is(are) skipped and a
warning is shown (\code{nonStop = TRUE}). Note that any other error
(such as xml import error) will abort import regardless of
parameter \code{nonStop}.
}
\author{
Michele Stravs \href{mailto:stravs@imsb.biol.ethz.ch}{stravs@imsb.biol.ethz.ch}
}
