% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser-utils.R
\name{specify}
\alias{specify}
\title{Specify a grammar term including test cases}
\usage{
specify(query, testcases, success = NULL, leftovers = NULL)
}
\arguments{
\item{query}{The Ramble parser to define}

\item{testcases}{The testcases to run through the parser, a list of strings}

\item{leftovers}{The expected leftovers, or NULL if the parsing should fail}
}
\value{
The parser.

The function exits with an error if a test fails.
If you specify any NULL results, \code{leftovers} needs to be a list, not vector!
}
\description{
Specify a grammar term including test cases
}
