% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser-utils.R
\name{warnifnot}
\alias{warnifnot}
\title{warnifnot(): a "only-warning" version of stopifnot():
{Yes, learn how to use do.call(substitute, ...) in a powerful manner !!}}
\usage{
warnifnot(..., exprs, exprObject, local = TRUE)
}
\description{
warnifnot(): a "only-warning" version of stopifnot():
{Yes, learn how to use do.call(substitute, ...) in a powerful manner !!}
}
