% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/MsBackendMapping.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{[,MsBackendMapping-method}
\alias{as.list,MsBackendMapping-method}
\alias{intensity,MsBackendMapping-method}
\alias{mz,MsBackendMapping-method}
\alias{lengths,MsBackendMapping-method}
\title{Internal page for hidden aliases}
\usage{
\S4method{as.list}{MsBackendMapping}(x)

\S4method{intensity}{MsBackendMapping}(object)

\S4method{mz}{MsBackendMapping}(object)

\S4method{lengths}{MsBackendMapping}(x, use.names = FALSE)
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
