% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/MsBackendMapping.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{[,MsBackendMapping-method}
\alias{backendMerge,MsBackendMapping-method}
\alias{.peaksData.MsBackendMapping}
\alias{peaksData,MsBackendMapping-method}
\alias{intensity,MsBackendMapping-method}
\alias{length,MsBackendMapping-method}
\alias{mz,MsBackendMapping-method}
\alias{lengths,MsBackendMapping-method}
\alias{export,MsBackendMapping-method}
\alias{spectraData<-,MsBackendMapping-method}
\alias{spectraData,MsBackendMapping-method}
\alias{spectraVariables,MsBackendMapping-method}
\alias{$,MsBackendMapping-method}
\alias{$<-,MsBackendMapping-method}
\alias{peaksData<-,MsBackendMapping-method}
\title{Internal page for hidden aliases}
\usage{
\S4method{backendMerge}{MsBackendMapping}(object, ...)

.peaksData.MsBackendMapping(object)

\S4method{peaksData}{MsBackendMapping}(object)

\S4method{intensity}{MsBackendMapping}(object)

\S4method{length}{MsBackendMapping}(x)

\S4method{mz}{MsBackendMapping}(object)

\S4method{lengths}{MsBackendMapping}(x, use.names = FALSE)

\S4method{export}{MsBackendMapping}(object, x, file = tempfile(), progress = FALSE, ...)

\S4method{spectraData}{MsBackendMapping}(object) <- value

\S4method{spectraData}{MsBackendMapping}(object, columns = object@spectraVariables)

\S4method{spectraVariables}{MsBackendMapping}(object)

\S4method{[}{MsBackendMapping}(x, i, j, ..., drop = FALSE)

\S4method{$}{MsBackendMapping}(x, name)

\S4method{$}{MsBackendMapping}(x, name) <- value

\S4method{peaksData}{MsBackendMapping}(object) <- value
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
